requirements.txt
altair==4.2.0
fiona==1.8.21
beautifulsoup4==4.10.0
pygeos==0.12.0
geopandas==0.10.2
gpdvega==0.1.1rc1
graphviz==0.19.2
ipywidgets==7.7.0
lxml==4.8.0
mapclassify==2.4.3
matplotlib==3.5.2
numpy==1.21.5
pandas==1.3.5
pillow==9.1.0
plotly==4.4.1
requests==2.27.1
scikit-learn==0.24.2
streamlit==1.8.1
seaborn==0.11.2
streamlit-folium==0.6.15